requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pymongo==4.5.0
click==8.1.7
requests==2.31.0
dnspython==2.4.2
gunicorn==21.2.0